requirements.txt
asgiref==3.7.2
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytz==2023.3
sqlparse==0.4.4
